Use sqlite for wordnet integration
This is a rewrite of the wordnet module to use the sqlite database from
http://wnsql.sourceforge.net . It largely maintains feature parity, but
the API has been changed to be more regular. The changes are roughly:

  get -> getWord, getWordById
  lookup -> findWords
  lookupSynonyms -> word.senses[n].getSynonymns

The file download functionality has been removed for the moment, as it's
not currently clear to me how that should work in the larger scheme of
things.