Fix "_ is not a function" error when webpacked (#554)
* Fix "_ is not a function" error when webpacked

In our application we use natural in lambda environment. To reduce size of bundle it is wrapped by webpack.
In such a case we are getting error `_ is not a function` on the line when all other dependencies work without issues. Not sure why is this happening but simple change to functional syntax that is being used across whole natural codebase fixes the issue.

* no new line